moved collisions to proper folder
