+++ play.py
@@ -1,7 +1,7 @@
 import pygame
 
-from collisions import Collisions
 from modules.behavior.camera import Camera
+from modules.behavior.collisions import Collisions
 from modules.objects.tiles import Ground
 from modules.state import GameState
 from modules.world.world import TILE_SIZE, WORLD_WIDTH
